backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
httpx
python-multipart
email-validator
